+++ tests/test_dataset.py
@@ -137,6 +137,11 @@ def test_insert(ds1):
     assert ds1.names == ["Foo", "Bar", "Baz"]
     assert ds1[-1] == da
 
+def test_insert_wrong_type(ds1):
+
+    with pytest.raises(ValueError):
+        ds1["Foo"] = np.zeros(ds1.shape)
+
 
 def test_insert_fail(ds1):
     da = ds1[0]
@@ -1601,3 +1606,8 @@ def test_interp_na():
     dsi = ds.interp_na(fill_value="extrapolate")
     assert dsi.Foo.to_numpy()[0] == pytest.approx(0.0)
     assert dsi.Bar.to_numpy()[2] == pytest.approx(4.0)
+
+def test_plot_scatter():
+    ds = mikeio.read("tests/testdata/oresund_sigma_z.dfsu", time=0)
+    ds.plot.scatter(x="Salinity", y="Temperature", title="S-vs-T")
+
